formatting + circular import fix
