+++ packages/insomnia-inso/src/errors.ts
@@ -1,0 +1,9 @@
+export class InsoError extends Error {
+  cause?: Error | null;
+
+  constructor(message: string, cause?: Error) {
+    super(message);
+    this.name = 'InsoError';
+    this.cause = cause;
+  }
+}
