+++ packages/insomnia-smoke-test/.gitignore
@@ -18,3 +18,5 @@ fixtures/inso-nedb/insomnia.CookieJar.db
 fixtures/inso-nedb/insomnia.GitRepository.db
 fixtures/inso-nedb/insomnia.OAuth2Token.db
 fixtures/inso-nedb/insomnia.PluginData.db
+fixtures/inso-nedb/insomnia.CaCertificate.db
+fixtures/inso-nedb/insomnia.CloudCredential.db
